BIG Update
In this updated the whole UI was remade from the ground. The old one
inspired by Yeezy.com was too limiting and ugly. I decided to go with a
UI style I'm more accostumed to, and I think the final result was better
because of this.

Now the album page is almost complete. The only things missing currently
are: complete mini player and each song's duration.

The main page was also remade and now should be better to contribute to.

In the future I will make the main page and the singles page get the
same info from the same file, as to make this part of the code also
better to contribute.

Also, I need to make the singles pages one single dynamic page just like
the albums page. Shouldn't be too hard as those pages basically already
work like that