Centralizing stuff
This update now makes it so that every file related to songs, or JSONs
about them, are in a single folder, making contributing and mantaining a
bit easier.

Also, now when you're in your phone, when you click on the mini player,
it opens a menu similar to Spotify. Right now, it's very simple, but
within the next few days/weeks everything will be fixed.

Also, now there's an about page and a page about how to contribute to
the project. I think it explains the basics, but it will be improved
alongside the rest of the project.