Project may enter alpha phase after this commit
This commit basically made the website functional. Now when you play any
song from an album, when the song ends it properly goes to the next one
in the list, and now the play/pause button at the top of the page
actually pauses the album instead of just restarting the playback.

Also, now the mobile version of the player is done. It uses the drawer
component from Shadcn and I think its the best it could be. In the
future, support for looping, volume and shuffle will be added. Lyrics
and song explanations will be added too but right now I think it's
better to add more songs for further testing.