Backend for album explanations done
All the backend is now done for the explanation of albums. Right now, it
only shows on desktop as the way I built the frontend will make it a bit
trickier to add support for phones. I will figure it out tomorrow.

Right now, all pages have "WIP" as their explanation, as I didn't feel
like going on wikis and summarizing their key points. This will however
be added in the future.

For the future, I will look into adding support for lyrics (maybe even
synced ones). I will first have to see if some service already has
lyrics for these unreleased projects, or if I will have to host them on
the website too.

Anyway, that's it for today.

Also, now the homepage is not "testing" anymore. Clear your cache and
enter the website and you will see the change in your URL bar.