Last ocmmit before going to bed
Added the Yeezus 2 songs, but still not the album tracklist. Fixed bugs
and now you can share what you're listening to, as the url bar now
contains the song's location. Just click on the share button and you're
good to go.

Also now "Only For The Night" serves as a interlude between Runaway and
Lonely. It's not like that in the original but I thought it would be a
better transition.