BIG update
Added Job Well Done as an album and started making the album
explanation.

Remade the testing homepage, which will futurely be the new homepage.
It's currently not done.

Removed the need to add the song location when making a JSON. It will be
based of its title. Be sure to put it correctly, including capitalized
letters.

General fixes.