Dynamic singles pages
Now pages for singles are dynamically generated, so a single folder is
required for making a new single. In this folder, thy shall have a JSON
file, its name being the name of the folder it's on, a lyrics.txt file
and an explanation.txt file.

Now, there's also a secret song that you can access. Try to find it
within my code.

Removed the folder Yeezus II inside the Yeezus 2 folder, and added
two new tracks to the Yeezus 2 tracklist.