Yeezus 2 update
Finally made the tracklist of Yeezus 2. It's not great, but I think it
gets the vibe check for now, probably going to change when the album
gets a album explanation update.

Did a rewrite of the function that skips songs. This rewrite aims to
make it easier to understand, and not that nested like before. Probably
will be a pain in the ass to change anything though, so I hope it works.

Also changed the tag size of the testing homepage, and didn't see how it
looked, so it may look like trash now.

Changed some song names but it's not that important.