Synced lyrics update
So now the project finally has support for synced lyrics! The file
format stays the same (although now you have to put the file name in all
lower case because it's easier for finding it and stuff).

Also, now at least half of the Bad Bitch Playbook has synced lyrics, and
I will probably add the rest in the next update. The songs are really
short so adding the synced lyrics is not that hard.