May God be with us & make this work wihout tests
Added Donda 2.
Changed the album page's track list to have a different UI and removed
the cover from the songs 'cause it's dumb to be that way.
Changed a little bit of the lyric's UI but only for PC 'cause I forgot
to add the smartphone UI and now it's too late to add it. Will add it
tomorrow.