MASSIVE update
This update adds lots of stuff, so that's why the project was kinda dead
for a little bit.

Firstly, now there is a settings page where you can change you theme and
your album page style.

Now the album page has two styles: the original one, and a new one based
on YT Music's UI. This new UI is kinda buggy, so it's not that
recommended to use.

Now the backend for the album page is also a little bit different. Now
it only contains the logic (or backend if you want to call it that) and
the actual UI is its own React Component. This change is important, as I
want to make the page a server component, for it to be more modular and
for it to have stuff like custom titles, and stuff like that. For that
to happen, I need the current page's code to be separated from the
page's code.

Also FINALLY added the album explanation for Yandhi, so right now I
would say it is about 90% done, just need to add more synced lyrics...

This update will probably introduce lots of bugs that I didn't catch
when testing, so expect a hotfix update soon...ish, I guess...

Yeah, that's it. About 4 days of coding for this. Around 7 hours of
total coding time, if my calculations are correct.