1/2
This commit is the first part of the last commit that I had to revert

This commit is mainly focused on UI changes and QoL features. For
example, if the song hasn't loaded yet, a spinning icon will show to
indicate that.

Now, the page's title will be of the album's or single's title too btw.

Aside from these changes, most stuff was already introduced in the
reverted commit so nothing to say really.

The next commit (aside the probable hotfix one) will be the
implementation of the credits that the reverted commit brought.

I decided to split these two because it would be easier to identify
possible problems that could happen.