New stuff to the settings page
More UI tweaks. Most importantly, if a song fails to load, a error will
show up in the mobile version (the desktop implementation was not done
yet 'cause I didn't have a clever way to show the error).

Also, now in the settings page you can choose the alignment of the
synced lyrics and normal lyrics individually.

Now the tutorial stuff also shows a tooltip telling the user to click on
the mini song controls thing to open the mini player. This should be the
last not user-intuitive thing in the site.