First steps
Added the logic behind the shuffle. Now a list of indexes will be made
when the page loads. Still need a way for every time the user skips a
song, the song that plays is the one from the array instead of the
normal one. Should be something like "if the number is 0 and the user
skipped, go to position 1", kinda like coordinates.

Right now having a really tough time doing this because of mental fog.
Will try to continue doing this tomorrow.