So Help Me God 1/2
This update adds several fixes and UI/UX improvements

The most important feature is the addition of a system where if an album
has multiple covers, the user can check them and set it as their default
album cover. I still need to add a description thing too but for now
this is enough. I don't wanna bloat this commit too much.

Next commit will fix more stuff and add the description stuff with the
addition of So Help Me God. I will also try to rewrite the legacy album
page, as it is right now disabled due to many problems with it.